+++ src/jslicense.page.ts
@@ -22,27 +22,42 @@ export default function* () {
         <tbody>
             <tr>
                 <td><a href="/js/register-sw.js">register-sw.js</a></td>
-                <td><a rel="license" href="https://spdx.org/licenses/AGPL-3.0-or-later.html">AGPL-3.0-or-later</a></td>
+                <td><a rel="license" href="https://www.gnu.org/licenses/agpl-3.0.html">AGPL-3.0-or-later</a></td>
                 <td><a href="/js/register-sw.js">Source</a></td>
             </tr>
             <tr>
                 <td><a href="/js/pagefind-init.js">pagefind-init.js</a></td>
-                <td><a rel="license" href="https://spdx.org/licenses/AGPL-3.0-or-later.html">AGPL-3.0-or-later</a></td>
+                <td><a rel="license" href="https://www.gnu.org/licenses/agpl-3.0.html">AGPL-3.0-or-later</a></td>
                 <td><a href="/js/pagefind-init.js">Source</a></td>
             </tr>
             <tr>
                 <td><a href="/js/contact-reveal.js">contact-reveal.js</a></td>
-                <td><a rel="license" href="https://spdx.org/licenses/AGPL-3.0-or-later.html">AGPL-3.0-or-later</a></td>
+                <td><a rel="license" href="https://www.gnu.org/licenses/agpl-3.0.html">AGPL-3.0-or-later</a></td>
                 <td><a href="/js/contact-reveal.js">Source</a></td>
             </tr>
             <tr>
                 <td><a href="/sw.js">sw.js</a></td>
-                <td><a rel="license" href="https://spdx.org/licenses/AGPL-3.0-or-later.html">AGPL-3.0-or-later</a></td>
+                <td><a rel="license" href="https://www.gnu.org/licenses/agpl-3.0.html">AGPL-3.0-or-later</a></td>
                 <td><a href="/sw.js">Source</a></td>
             </tr>
             <tr>
                 <td><a href="/pagefind/pagefind-ui.js">pagefind-ui.js</a></td>
-                <td><a rel="license" href="https://opensource.org/licenses/MIT">MIT</a></td>
+                <td><a rel="license" href="https://www.gnu.org/licenses/license-list.html#Expat">Expat (MIT)</a></td>
+                <td><a href="https://github.com/pagefind/pagefind">Upstream source</a></td>
+            </tr>
+            <tr>
+                <td><a href="/pagefind/pagefind.js">pagefind.js</a></td>
+                <td><a rel="license" href="https://www.gnu.org/licenses/license-list.html#Expat">Expat (MIT)</a></td>
+                <td><a href="https://github.com/pagefind/pagefind">Upstream source</a></td>
+            </tr>
+            <tr>
+                <td><a href="/pagefind/pagefind-highlight.js">pagefind-highlight.js</a></td>
+                <td><a rel="license" href="https://www.gnu.org/licenses/license-list.html#Expat">Expat (MIT)</a></td>
+                <td><a href="https://github.com/pagefind/pagefind">Upstream source</a></td>
+            </tr>
+            <tr>
+                <td><a href="/pagefind/pagefind-modular-ui.js">pagefind-modular-ui.js</a></td>
+                <td><a rel="license" href="https://www.gnu.org/licenses/license-list.html#Expat">Expat (MIT)</a></td>
                 <td><a href="https://github.com/pagefind/pagefind">Upstream source</a></td>
             </tr>
         </tbody>